+++ components/index.tsx
@@ -1,5 +1,5 @@
 /* @remove-on-es-build-begin */
-import './alphaWarning';
+import './_util/alphaWarning';
 
 // this file is not used if use https://github.com/ant-design/babel-plugin-import
 const ENV = process.env.NODE_ENV;
